requirements.txt
streamlit==0.79.0
qrcode==7.3
Pillow==8.2.0
